+++ tests/Controller/InformationCollectionLegacyControllerTest.php
@@ -6,6 +6,7 @@ use eZ\Publish\Core\MVC\Symfony\Controller\Content\ViewController;
 use eZ\Publish\Core\Repository\LocationService;
 use eZ\Publish\Core\Repository\Values\Content\Location;
 use Netgen\Bundle\EzFormsBundle\Form\DataWrapper;
+use Netgen\Bundle\EzFormsBundle\Form\Payload\InformationCollectionStruct;
 use Netgen\Bundle\InformationCollectionBundle\Form\Builder\FormBuilder;
 use PHPUnit\Framework\TestCase;
 use Netgen\Bundle\InformationCollectionBundle\Controller\InformationCollectionLegacyController;
@@ -155,9 +156,9 @@ class InformationCollectionLegacyControllerTest extends TestCase
             ->method('isValid')
             ->willReturn(true);
 
-        $this->form->expects($this->once())
+        $this->form->expects($this->exactly(2))
             ->method('getData')
-            ->willReturn(new DataWrapper('payload'));
+            ->willReturn(new DataWrapper(new InformationCollectionStruct()));
 
         $this->form->expects($this->once())
             ->method('createView');
@@ -207,8 +208,9 @@ class InformationCollectionLegacyControllerTest extends TestCase
             ->method('isValid')
             ->willReturn(false);
 
-        $this->form->expects($this->never())
-            ->method('getData');
+        $this->form->expects($this->exactly(1))
+            ->method('getData')
+            ->willReturn(new DataWrapper(new InformationCollectionStruct()));
 
         $this->form->expects($this->once())
             ->method('createView');